Arquivo HTML removido, docgen com argumentos.

+++ docgen.py
@@ -1,9 +1,7 @@
 from os import listdir, getcwd
-from os.path import isfile, join, splitext, basename 
-
-HTML_FILE = "Principal.html"
-PAPER_PATH = "Artigos/"
-DESC_PATH = "Descricoes/"
+import webbrowser
+import sys
+from os.path import isfile, join, splitext, basename, exists
 
 def html_header():
     html = open(HTML_FILE,"w+")
@@ -42,9 +40,39 @@ def html_table_gen(file_names):
         html = open(HTML_FILE,"a")
         html.write("\n\t\t<tr><td><a href=\""+PAPER_PATH+fn+".pdf"+"\">"+title+"</a></td><td>"+content+"</td></tr>")
 
-paper_list = [splitext(f)[0] for f in listdir(PAPER_PATH) if isfile(join(PAPER_PATH, f))]
-
 try:
+    open_browser = False
+    arg_path = '.'
+
+    # Valida parametros passados na chamada
+    if len(sys.argv)<1 and len(sys.argv)>3:
+        print("É necessário especificar apenas um diretório válido!")
+        raise Exception
+
+    for arg in sys.argv:
+        if arg=="-o":
+            open_browser=True
+        else:
+            arg_path=arg
+
+    path_referencia = getcwd()+"/"+arg_path
+    if path_referencia[-1]!='/':
+        path_referencia=path_referencia+'/'
+    HTML_FILE = path_referencia+"Principal.html"
+    PAPER_PATH = path_referencia+"Artigos/"
+    DESC_PATH = path_referencia+"Descricoes/"
+   
+    # Verifica se existem as pastas base Artigos e Descricoes    
+    if not(exists(PAPER_PATH)):
+        print("Não existe um repositório para Artigos!")
+        raise Exception 
+    if not(exists(DESC_PATH)):
+        print("Não existe um repositório para Descricoes!")
+        raise Exception 
+
+    # Gera a lista de artigos de acordo com os arquivos na pasta de papers
+    paper_list = [splitext(f)[0] for f in listdir(PAPER_PATH) if isfile(join(PAPER_PATH, f))]
+
     for paper_file in paper_list:
         if not(isfile(DESC_PATH+paper_file+".txt")):
            f= open(DESC_PATH+paper_file+".txt","w+")
@@ -57,5 +85,10 @@ try:
     html_table_gen(paper_list)
     html_footer()
     print("** Documentação criada com sucesso **")
+
+
+    if open_browser:
+        webbrowser.open_new_tab(HTML_FILE)
+
 except:
-    print("Erro ao criar documentação!")
+    print("Documentação não gerada!")
